requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
finbert-embedding>=0.1.4
tqdm>=4.65.0
optuna>=3.2.0